// Homework3/etnchen/src/types.ts
// Global types and interfaces are stored here.
export interface Margin {
    readonly left: number;
    readonly right: number;
    readonly top: number;
    readonly bottom: number;
}

export interface ComponentSize {
    width: number;
    height: number;
}

export interface Point {
    readonly posX: number;
    readonly posY: number;
}

export interface Bar{
    readonly total: number;
}

// export interface PieChart{
//     readonly event: string;
//     readonly sport: string;
// }